Implementing setting of burn speed.
WARNING: my dvd test burn failed if a speed was set!! CD was ok though.

Use "Max" as speed if that is the case for you to get the old behavior,
and _please_ report this as a bug! Please include the console output if
debug is enabled, or otherwise some info on the burner that was used.


(Old svn revision: 4972)